+++ NEWS.md
@@ -1,4 +1,11 @@
 # RFishBC 0.1.9000 ongoing
+* Added several items (mostly images) to .Rbuildignore to lower bloat of package.
+* Added .Rbuildignore to remove sticker folder from package directory.
+* Added validation tests (still need to be run manually) from Sullivan data.
+* `combineData()`: Modified. Changed `outFormat=` to `formatOut=`. Added `outType=` to allow user to choose either radial or incremental measurements as output.
+* `digitizeRadii()`: Modified. Changed `showTransect=` to `makeTransect`. Changed so that `snap2Transect` is changed to `FALSE` if it is `TRUE` and `makeTransect=FALSE`. Radii are now computed as the cumulative sum of increments rather than the distance from the selected point back to the structure center. This addresses [#32](https://github.com/droglenc/RFishBC/issues/32).
+* `RFBCoptions()`: Modified. Now includes `makeTransect=` used in `digitizeRadii()`. Now includes `connect=`, `col.connect=`, and `lwd.connect=` used in `showDigitizedImage()`. Because of those changes `showTransect=` has been removed.
+* `showDigitizedImage()`: Modified. Changed `showTransect=` to `connect=`, `col.transect=`  to `col.connect=`, and `lwd.transect=` to `lwd.connect=` (this helps address [#32](https://github.com/droglenc/RFishBC/issues/32)). Also removed the points that are plotted at the structure center and, if not an annulus, margin (this addresses [#31](https://github.com/droglenc/RFishBC/issues/31)). Made more robust checks for when the user tries to plot two images from different structures.
 
 # RFishBC 0.1.2 9-Dec-18
 * Added dependency to `cli` package for `DONE()`, `NOTE()`, and `RULE()` (see below).